Уменьшено время атаки оружия. F Исправлен расчет времени атаки.

+++ HeroInBattle.cs
@@ -170,7 +170,7 @@ namespace Fantasy_King_s_Battle
 
         private int TimeAttack()
         {
-            int timeAttack = PlayerHero.Weapon.Item.TimeHit / 100 * Config.STEPS_IN_SECOND;
+            int timeAttack = (int)(PlayerHero.Weapon.Item.TimeHit / 100.00 * Config.STEPS_IN_SECOND);
 
             Debug.Assert(timeAttack > 0);
